backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "REST API для мобильного приложения обмена книгами"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[all]>=0.117.1",
    "sqlalchemy>=2.0.43",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "pillow>=10.1.0",
    "httpx>=0.25.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.24.0",
    "email-validator>=2.1.0",
    "phonenumbers>=8.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
